+++ gulpfile.js/task-config.mjs
@@ -1,8 +1,7 @@
-import pathConfig from "./path-config.json" with { type: "json" };
-
 export default {
   html: true,
   images: true,
+  cloudflare: false,
   cloudinary: false,
   fonts: true,
   static: true,
@@ -10,23 +9,9 @@ export default {
   stylesheets: true,
   esbuild: true,
 
-  browserSync: {
-    server: {
-      baseDir: pathConfig.dest
-    }
-  },
-
-  workboxBuild: {
-    globDirectory: pathConfig.dest,
-    globPatterns: ["**/*.{html,json,js,css,png,jpg,gif,svg}"],
-    // Create service-worker.js source file and define `swSrc` to use `injectManifest` method
-    // swSrc: `${pathConfig.src}/sw.js`,
-    swDest: `${pathConfig.dest}/sw.prod.js`
-  },
-
   production: {
     rev: {
-      exclude: ["favicon.ico", "robots.txt", "_headers", "_redirects"]
+      exclude: ["_headers", "_redirects"]
     }
   },