+++ gulpfile.js/tasks/cloudinary.mjs
@@ -30,10 +30,10 @@ export class CloudinaryRegistry extends DefaultRegistry {
       dest: projectPath(pathConfig.src, pathConfig.data?.src ?? ""),
       manifest: config.manifest
         ? projectPath(
-            pathConfig.src,
-            pathConfig.data?.src ?? "",
-            config.manifest
-          )
+          pathConfig.src,
+          pathConfig.data?.src ?? "",
+          config.manifest
+        )
         : null
     };
   }
@@ -58,7 +58,7 @@ export class CloudinaryRegistry extends DefaultRegistry {
     }
 
     const cloudinaryTask = () =>
-      src(path.join(paths.src, "**", `*.{${this.config.extensions}}`), {
+      src(path.join(paths.src, "**", this.config.extensions.length > 1 ? `*.{${this.config.extensions}}` : `*.${this.config.extensions}`), {
         encoding: false,
         since: lastRun(cloudinaryTask)
       })