+++ src/data/global.mjs
@@ -1,0 +1,6 @@
+export default {
+  meta: {
+    lang: "en",
+    title: "Blendid!"
+  }
+};